LP#1879983: Release notes for Curbside Pickup
This is in the first patch in a series for a new feature in
Evergreen called Curbside Pickup.

The Curbside Pickup features in Evergreen provides an interface to help
facilitate contact free pickup of library materials.  It provides a dedicated
interface in the staff client for library staff to track and manage curbside
pickup appointments and materials through the various stages of the process.
Staff can also schedule pickup appointments on behalf of patrons.  This feature
also allows patrons to schedule their own curbside pickup appointments in their
OPAC account, as well as inform the library when they have arrived and are
waiting for their materials.

To test
-------
(Testing notes were contributed by Andrea Neiman)

In order to test Curbside, you will need to place and capture holds at the
respective location you’re testing.  This work does not interfere with existing
hold rules or parameters, except to force checkout override in certain
circumstances.

In addition, the circ.curbside library setting must be enabled. These testing
suggestions assume that the A/T event definitions included in this patch
series have been enabled.

Things to test:

* MyAccount correctly reflects scheduling/scheduled options
* Alert staff on my arrival should move batch to the Patron is Outside
  tab
* Update appointment/cancel appointment options should move or remove
  the batch appropriately
* If a patron has pickups at multiple locations, they should be able
  to schedule separate pickups per location
* Email notices at various steps of the way are firing correctly
* Patron should get email for hold pickup, including scheduling
  curbside.
* Patron should get email confirming curbside pickup
* Patron should get email with checked out items, if email receipts
  are enabled
* Correct “batching” of items for pickup
* Batch moving correctly through tabs
* Batch is physically collocated - i.e. not containing pickups for
  other OUs
* Checkouts are performed correctly
* Check against patrons with blocks - force override should be in effect
* Settings are being respected with regard to pickup scheduling & time
  constraints
* Can’t create pickups in the past
* Pickups respect library open hours as set in Org Unit Administration
* Can’t create pickups if the 15-minute block is filled

Sponsored-by: PaILS

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Michele Morgan <mmorgan@noblenet.org>