LP#1857150: eg-fm-record-editor: handle unavailable linked values
In some cases, a record editor might be handed a base object
that has a linked column whose values cannot be retrieved
because the user has permission to fetch the base object
but not the linked objects. This patch makes the record editor
more resilient in the face of that situation.

Sponsored-by: Evergreen Community Development Initiative

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Ruth Frasur <rfrasur@library.in.gov>
Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Jason Etheridge <jason@EquinoxInitiative.org>