LP#1979570: Require users to update old passwords
API call to get the age of a patron's password, this is displayed in the
user editor and myopac home page. Added a configurable variable in
config to set when the password reminder should show up. Patrons can not enter a duplicate password to bypass message. Passive Hook for
password update based on edit date. Example event
definition for password update notice.