LP#2136881: Detect requested, unused exact-match dynamic filters
Today, a search will fail if the user requests an exact match on a
record attribute filter (a dynamic filter, IOW).  This is because we
assume that user-entered values will be covered by the data that has
been indexed.  However, if none of the user-requested values exist in
the data, the correct answer is "no results" rather than an error.

This patch intends to do that by emitting a FALSE WHERE clause condition
to be used with that branch of the search, which will, in the end, cause
the search to succeed with an empty result.

Signed-off-by: Mike Rylander <mrylander@gmail.com>